Small update for Update.md (#222)
Me being a C n00b I tried adding a module and updating my core with the proposed general way first (without re-running cmake) and my module was not being detected and was going back and forth checking what i did wrong or if I put the module incorrectly or if the module was still compatible with this version of AC, went back to the full instructions and saw that maybe i needed to rerun the CMake part, after doing that I saw my module being detected when rebuilding and running my core now. If this line was here it would've save me a bit time and frustration figuring it out.