ac-dashboard-core-install: Fix conf file section
The current worldserver and authserver fails to load if you don't have the worldserver.conf and authserver.conf. The ./acore run-{world,auth}server commands also gets stuck in an infinite loop and bails (nice check!) due to worldserver and authserver returning a non-zero exit value.

This was required to make this work under Ubuntu on WSL2 running on Windows 11.