Update database-installation.md (#1011)
The instructions say 'To edit the automatic database updater you will find the necessary settings in authserver.conf and worldserver.conf under UPDATE SETTINGS.' referencing authserver.conf and worldserver.conf.

But then the next 2 sentences say to 'Start authserver' and 'Start worldserver'. It isn't clear that it's referring to a different authserver/worldserver than in the previous sentence.